Move some of add_assets code to a new module: url_selector
